requirements.txt
# Core dependencies
google-generativeai>=0.3.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Utilities
requests>=2.28.0
typing-extensions>=4.5.0 